- Se eliminó el antiguo archivo db.ts, que contenía ofertas y recomendaciones de productos obsoletas. - Se actualizaron las páginas index.astro y nosotros.astro para reflejar la nueva estructura de datos e importaciones. - Se mejoraron los detalles de los miembros del equipo para mayor claridad y se actualizó la sección de cultura empresarial para incorporar un nuevo componente.
