+++ src/components/home/fiestas-patrias-ofertas.astro
@@ -4,7 +4,7 @@ import { Button } from '@/components/ui/button'
 import { Card, CardContent } from '@/components/ui/card'
 import { Badge } from '@/components/ui/badge'
 import Counter from './Counter'
-import { patriasOffers } from '@/constants/db'
+import { patriasOffers } from '@/constants/db_product'
 ---
 
 <section class="py-16 bg-card relative overflow-hidden">
