--- src/components/product/data.ts
@@ -1,392 +1,0 @@
-export const laptops = [
-	{
-		name: "Laptop Asus Rog Strix Intel Core i9-14900Hx, 16GB, 1Tb, 16'' FHD,Video RTX 4060, Win 11 Pro",
-		price: 'S/ 1,469',
-		originalPrice: 'S/ 1,699',
-		brand: 'Asus',
-		rating: 4,
-		reviews: 128,
-		discount: 15,
-		stock: 5,
-		freeShipping: true,
-		specs: ['Intel Core i9', '16GB RAM', '1TB SSD', 'RTX 4060'],
-		installments: '12x S/ 122.42 sin interés',
-		image: '/productos/header_ofert1.png?height=300&width=300'
-	},
-	{
-		name: "Gamer HP Intel Core i7 RTX 3050 16GB 512 GB SSD Victus 12va Gen 15.6'' 15-fa0007la W11",
-		price: 'S/ 1,499',
-		originalPrice: 'S/ 1,599',
-		brand: 'HP',
-		rating: 5,
-		reviews: 89,
-		discount: 8,
-		stock: 12,
-		freeShipping: true,
-		isBestSeller: true,
-		specs: ['Intel Core i7', '16GB RAM', '512GB SSD', 'RTX 3050'],
-		installments: '18x S/ 81.61 sin interés',
-		image: '/productos/header_ofert2.png?height=300&width=300'
-	},
-	{
-		name: 'LAPTOP GAMER ACER NITRO V15 ANV15-51-711U i7 13° GEN. 15.6 FHD IPS 144Hz 16GB 512GB RTX4050 6GB W11',
-		price: 'S/ 1,549',
-		originalPrice: 'S/ 1,549',
-		brand: 'Acer',
-		rating: 4,
-		reviews: 67,
-		discount: 5,
-		stock: 8,
-		freeShipping: true,
-		specs: ['Intel Core i7', '16GB RAM', '512GB SSD', 'RTX 4050'],
-		installments: '24x S/ 61.21 sin interés',
-		image: '/productos/header_ofert3.png?height=300&width=300'
-	},
-	{
-		name: 'ACER Nitro V15 ANV15-51-56GH Ci5 13420H/ 8Gb Ram/ 512GB SSD/ RTX 3050 6GB/ 15.6”FHD W.11 - CSI',
-		price: 'S/ 1,469',
-		originalPrice: 'S/ 1,799',
-		brand: 'Acer',
-		rating: 5,
-		reviews: 156,
-		discount: 18,
-		stock: 3,
-		freeShipping: true,
-		isNew: true,
-		specs: ['Intel Core i5', '8GB RAM', '512GB SSD', 'RTX 3050'],
-		installments: '6x S/ 244.83 sin interés',
-		image: '/productos/header_ofert4.png?height=300&width=300'
-	}
-]
-
-export const audifonos = [
-	{
-		name: 'Audifono Redragon Zeus X PRO RGB H510-PRO-WL Wireless BLACK',
-		price: 's/ 89.99',
-		originalPrice: 's/ 119.99',
-		brand: 'Redragon',
-		rating: 5,
-		reviews: 234,
-		discount: 25,
-		stock: 15,
-		freeShipping: true,
-		isBestSeller: true,
-		specs: ['Bluetooth 5.0', 'RGB lighting', '40-hour battery life'],
-		installments: '3x s/ 29.99 sin interés',
-		image: '/productos/audifonos/audifono1.png?height=200&width=200'
-	},
-	{
-		name: 'Audífonos Bluetooth Haylou S30 PRO ANC Hybrid Hi-Res - Blanco',
-		price: 's/ 179.99',
-		originalPrice: 's/ 199.99',
-		brand: 'Haylou',
-		rating: 4,
-		reviews: 167,
-		discount: 10,
-		stock: 8,
-		freeShipping: true,
-		specs: ['Active Noise Cancellation', 'Hi-Res Audio', 'Touch controls'],
-		installments: '6x s/ 29.99 sin interés',
-		image: '/productos/audifonos/audifono2.png?height=200&width=200'
-	},
-	{
-		name: 'Audífonos Bluetooth Haylou S30 PRO ANC Hybrid Hi-Res - Negro',
-		price: 's/ 129.99',
-		originalPrice: 's/ 149.99',
-		brand: 'Haylou',
-		rating: 4,
-		reviews: 198,
-		discount: 13,
-		stock: 12,
-		freeShipping: true,
-		specs: ['Bluetooth 5.1', 'Low latency', 'Dual microphone'],
-		installments: '4x s/ 32.49 sin interés',
-		image: '/productos/audifonos/audifono3.png?height=200&width=200'
-	},
-	{
-		name: 'AUDIFONOS GAMER COMBAT NEGROAZUL',
-		price: 's/ 149.99',
-		originalPrice: 's/ 169.99',
-		brand: 'Combat',
-		rating: 5,
-		reviews: 145,
-		discount: 12,
-		stock: 6,
-		freeShipping: true,
-		isNew: true,
-		specs: ['50mm drivers', 'Noise isolation', 'Adjustable headband'],
-		installments: '5x s/ 29.99 sin interés',
-		image: '/productos/audifonos/audifono4.png?height=200&width=200'
-	},
-	{
-		name: 'Corsair HS80 RGB Wireless',
-		price: 's/ 149.99',
-		originalPrice: 's/ 179.99',
-		brand: 'Corsair',
-		rating: 4,
-		reviews: 112,
-		discount: 17,
-		stock: 9,
-		freeShipping: true,
-		specs: ['Dolby Atmos', 'Custom-tuned 50mm drivers', 'Up to 20 hours battery life'],
-		installments: '6x s/ 24.99 sin interés',
-		image: '/productos/audifonos/audifono5.png?height=200&width=200'
-	},
-	{
-		name: 'Audífonos Bluetooth Haylou S30 PRO ANC Hybrid Hi-Res',
-		price: 's/ 199.99',
-		originalPrice: 's/ 229.99',
-		brand: 'Haylou',
-		rating: 5,
-		reviews: 78,
-		discount: 13,
-		stock: 4,
-		freeShipping: true,
-		specs: ['Bluetooth 5.2', 'IPX5 water resistance', 'Touch controls'],
-		installments: '8x s/ 24.99 sin interés',
-		image: '/productos/audifonos/audifono6.png?height=200&width=200'
-	}
-]
-
-export const televisores = [
-	{
-		name: 'Samsung OLED 55" QN90A Neo QLED 4K',
-		price: 's/ 1,299.99',
-		originalPrice: 's/ 1,499.99',
-		brand: 'Samsung',
-		rating: 5,
-		reviews: 89,
-		discount: 13,
-		stock: 7,
-		freeShipping: true,
-		isBestSeller: true,
-		specs: ['Neo Quantum Processor 4K', 'Quantum HDR 32X', '120Hz refresh rate'],
-		installments: '12x s/ 108.33 sin interés',
-		image: '/productos/televisiones/tv1.png?height=200&width=200'
-	},
-	{
-		name: 'LG OLED 65" C1 Series 4K Smart TV',
-		price: 's/ 1,799.99',
-		originalPrice: 's/ 1,999.99',
-		brand: 'LG',
-		rating: 5,
-		reviews: 156,
-		discount: 10,
-		stock: 5,
-		freeShipping: true,
-		specs: ['α9 Gen4 AI Processor', 'Dolby Vision IQ', 'HDMI 2.1'],
-		installments: '18x s/ 99.99 sin interés',
-		image: '/productos/televisiones/tv2.png?height=200&width=200'
-	},
-	{
-		name: 'Sony OLED 55" A80J Bravia XR',
-		price: 's/ 1,399.99',
-		originalPrice: 's/ 1,599.99',
-		brand: 'Sony',
-		rating: 4,
-		reviews: 134,
-		discount: 12,
-		stock: 8,
-		freeShipping: true,
-		isNew: true,
-		specs: ['Cognitive Processor XR', 'XR OLED Contrast', 'Google TV'],
-		installments: '15x s/ 93.33 sin interés',
-		image: '/productos/televisiones/tv3.png?height=200&width=200'
-	},
-	{
-		name: 'TCL QLED 50" C825 Mini LED 4K',
-		price: 's/ 899.99',
-		originalPrice: 's/ 1,099.99',
-		brand: 'TCL',
-		rating: 4,
-		reviews: 67,
-		discount: 18,
-		stock: 12,
-		freeShipping: true,
-		specs: ['Quantum Dot technology', 'Mini LED backlight', '120Hz VRR'],
-		installments: '9x s/ 99.99 sin interés',
-		image: '/productos/televisiones/tv4.png?height=200&width=200'
-	},
-	{
-		name: 'samsung QLED 55" Q80A 4K Smart TV',
-		price: 's/ 999.99',
-		originalPrice: 's/ 1,199.99',
-		brand: 'Samsung',
-		rating: 4,
-		reviews: 98,
-		discount: 17,
-		stock: 10,
-		freeShipping: true,
-		specs: ['Quantum Dot Wide Color', 'Full Array Local Dimming', 'Dolby Vision'],
-		installments: '10x s/ 99.99 sin interés',
-		image: '/productos/televisiones/tv5.png?height=200&width=200'
-	},
-	{
-		name: 'samsung QLED 65" QN85A 4K Smart TV',
-		price: 's/ 1,199.99',
-		originalPrice: 's/ 1,399.99',
-		brand: 'Samsung',
-		rating: 5,
-		reviews: 45,
-		discount: 14,
-		stock: 6,
-		freeShipping: true,
-		specs: ['P5 AI Perfect Picture', '3-sided Ambilight', 'Android TV'],
-		installments: '12x s/ 99.99 sin interés',
-		image: '/productos/televisiones/tv6.png?height=200&width=200'
-	}
-]
-
-export const Sillas = [
-	{
-		name: 'Silla Gamer Reclinación 155° con Posapies y Masajeador',
-		price: 's/ 429.99',
-		originalPrice: 's/ 449.99',
-		brand: 'Codari',
-		rating: 5,
-		reviews: 267,
-		discount: 4,
-		stock: 15,
-		freeShipping: true,
-		isBestSeller: true,
-		specs: ['Reclinación 155°', 'Posapies ajustable', 'Masajeador incorporado'],
-		installments: '12x s/ 35.83 sin interés',
-		image: '/productos/sillas/silla1.png?height=200&width=200'
-	},
-	{
-		name: 'Silla Gamer Reclinación 155° con Posapies y Masageador',
-		price: 's/ 379.99',
-		originalPrice: 's/ 429.99',
-		brand: 'Codari',
-		rating: 4,
-		reviews: 189,
-		discount: 12,
-		stock: 11,
-		freeShipping: true,
-		specs: ['Reclinación 155°', 'Posapies ajustable', 'Diseño ergonómico'],
-		installments: '10x s/ 37.99 sin interés',
-		image: '/productos/sillas/silla2.png?height=200&width=200'
-	},
-	{
-		name: 'Silla Gamer Kuzler Con Masajeador Lumbar Color Rosa RIK-101R',
-		price: 's/ 449.99',
-		originalPrice: 's/ 499.99',
-		brand: 'Kuzler',
-		rating: 5,
-		reviews: 123,
-		discount: 10,
-		stock: 8,
-		freeShipping: true,
-		isNew: true,
-		specs: ['Reclinación 155°', 'Posapies ajustable', 'Masajeador lumbar'],
-		installments: '15x s/ 29.99 sin interés',
-		image: '/productos/sillas/silla3.png?height=200&width=200'
-	},
-	{
-		name: 'Silla Gamer Kuzler Negra con masajeador lumbar reposapiés y base metálica',
-		price: 's/ 199.99',
-		originalPrice: 's/ 249.99',
-		brand: 'Kuzler',
-		rating: 4,
-		reviews: 234,
-		discount: 20,
-		stock: 18,
-		freeShipping: true,
-		specs: ['Reclinación 155°', 'Posapies ajustable', 'Base metálica resistente'],
-		installments: '6x s/ 33.33 sin interés',
-		image: '/productos/sillas/silla4.png?height=200&width=200'
-	},
-	{
-		name: 'Silla Gamer Extingtion Sapp Gamuza-Cuerina Color Blanco',
-		price: 's/ 279.99',
-		originalPrice: 's/ 329.99',
-		brand: 'Extingtion',
-		rating: 4,
-		reviews: 156,
-		discount: 15,
-		stock: 13,
-		freeShipping: true,
-		specs: ['Reclinación 155°', 'Posapies ajustable', 'Diseño moderno'],
-		installments: '8x s/ 34.99 sin interés',
-		image: '/productos/sillas/silla5.png?height=200&width=200'
-	},
-	{
-		name: 'Silla Gamer Con Luces RGB Titan Con Masajeador Reposapiés 135° Blanco',
-		price: 's/ 199.99',
-		originalPrice: 's/ 229.99',
-		brand: 'Titan',
-		rating: 4,
-		reviews: 178,
-		discount: 13,
-		stock: 16,
-		freeShipping: true,
-		specs: ['Reclinación 135°', 'Luces RGB personalizables', 'Masajeador lumbar'],
-		installments: '7x s/ 28.57 sin interés',
-		image: '/productos/sillas/silla6.png?height=200&width=200'
-	}
-]
-
-// Ofertas Flash
-export const flashDeals = [
-	{
-		name: 'Laptop Lenovo Legion 5, Amd Ryzen 5 5600H, 8GB DDR4, SSD 512Gb, 15.6" FHD ',
-		price: 'S/ 2500',
-		originalPrice: 'S/ 3000',
-		brand: 'Lenovo',
-		rating: 5,
-		reviews: 345,
-		discount: 17,
-		stock: 3,
-		freeShipping: true,
-		timeLeft: '2h 15m',
-		specs: ['M1 chip', '10.9-inch display', '12MP cameras'],
-		image: '/productos/ofert1.png?height=200&width=200',
-		installments: '12x S/ 208.33 sin interés'
-	},
-	{
-		name: 'Teclado Gamer Marca Logitech G213 Prodigy Con Iluminación RGB, Black, Alámbrico - USB',
-		price: 'S/ 222.99',
-		originalPrice: 'S/ 249.99',
-		brand: 'Logitech',
-		rating: 5,
-		reviews: 189,
-		discount: 17,
-		stock: 2,
-		freeShipping: true,
-		timeLeft: '1h 45m',
-		specs: ['M2 chip', '13.6-inch display', '18-hour battery'],
-		image: '/productos/ofert2.png?height=200&width=200',
-		installments: '12x S/ 208.33 sin interés'
-	},
-	{
-		name: 'Case Gamer Halion Aries 06 Con Fuente de 600W, 4 LED-ARGB, Vidrio Templado, Black',
-		price: 'S/ 799.99',
-		originalPrice: 'S/ 899.99',
-		brand: 'Halion',
-		rating: 5,
-		reviews: 567,
-		discount: 12,
-		stock: 1,
-		freeShipping: true,
-		timeLeft: '3h 30m',
-		specs: ['A16 Bionic chip', '6.1-inch display', '48MP camera'],
-		image: '/productos/ofert3.png?height=200&width=200',
-		installments: '12x S/ 208.33 sin interés'
-	}
-]
-
-export const categories = [
-	{ title: 'Laptops Gaming', icon: '🎮', productCount: 156, image: '/productos/categorias/cat1.png?height=100&width=100' },
-	{ title: 'Smartphones', icon: '📱', productCount: 234, image: '/productos/categorias/cat2.png?height=100&width=100' },
-	{ title: 'Audífonos', icon: '🎧', productCount: 89, image: '/productos/categorias/cat3.png?height=100&width=100' },
-	{ title: 'Televisores', icon: '📺', productCount: 67, image: '/productos/categorias/cat4.png?height=100&width=100' },
-	{ title: 'Smartwatches', icon: '⌚', productCount: 45, image: '/productos/categorias/cat5.png?height=100&width=100' },
-	{ title: 'Sillas', icon: '🔌', productCount: 178, image: '/productos/categorias/cat6.png?height=100&width=100' }
-]
-
-export const brandLogos = [
-	{ name: 'Logitech', src: '/productos/marcas/logo1.png' },
-	{ name: 'Samsung', src: '/productos/marcas/logo2.png' },
-	{ name: 'HyperX', src: '/productos/marcas/logo3.png' },
-	{ name: 'Asus', src: '/productos/marcas/logo4.png' }
-]
