--- src/constants/db.ts
@@ -1,216 +1,0 @@
-export const patriasOffers = [
-	{
-		name: 'iPhone 15 Pro Max 256GB',
-		category: 'CELULARES',
-		price: 999,
-		originalPrice: 1299,
-		image: '/productos/patrias/iPhone-15-Pro-Max-256GB.png',
-		discount: 23,
-		brand: 'Apple',
-		rating: 5,
-		reviews: 234,
-		stock: 5,
-		isLimitedTime: true
-	},
-	{
-		name: 'MacBook Air M2 512GB',
-		category: 'LAPTOPS',
-		price: 1199,
-		originalPrice: 1499,
-		image: '/productos/patrias/MacBook-Air-M2-512GB.png',
-		discount: 20,
-		brand: 'Apple',
-		rating: 5,
-		reviews: 189,
-		stock: 3,
-		isLimitedTime: true
-	},
-	{
-		name: 'Samsung Galaxy S24 Ultra',
-		category: 'CELULARES',
-		price: 899,
-		originalPrice: 1199,
-		image: '/productos/patrias/Samsung-Galaxy-S24-Ultra.png',
-		discount: 25,
-		brand: 'Samsung',
-		rating: 4,
-		reviews: 156,
-		stock: 8,
-		isLimitedTime: true
-	},
-	{
-		name: 'Sony WH-1000XM5 Headphones',
-		category: 'AUDIO',
-		price: 299,
-		originalPrice: 399,
-		image: '/productos/patrias/Sony-WH-1000XM5-Headphones.png',
-		discount: 25,
-		brand: 'Sony',
-		rating: 5,
-		reviews: 298,
-		stock: 12,
-		isLimitedTime: true
-	}
-]
-
-export const products = [
-	{
-		name: 'ASUS ROG Zephyrus G14 14"',
-		category: 'LAPTOPS',
-		price: 5000,
-		originalPrice: 5099,
-		image: '/productos/header_ofert1.png?height=300&width=300',
-		discount: 2,
-		brand: 'ASUS',
-		rating: 5,
-		reviews: 234,
-		stock: 5,
-		isLimitedTime: true
-	},
-	{
-		name: 'VICTUS Gaming 15.6 FHD',
-		category: 'LAPTOPS',
-		price: 5400,
-		originalPrice: 5700,
-		image: '/productos/header_ofert2.png?height=300&width=300',
-		discount: 5,
-		brand: 'HP',
-		rating: 4,
-		reviews: 176,
-		stock: 6,
-		isLimitedTime: true
-	},
-	{
-		name: 'NITRO 5 Gaming 15.6 FHD',
-		category: 'LAPTOPS',
-		price: 4000,
-		originalPrice: 4500,
-		image: '/productos/header_ofert3.png?height=300&width=300',
-		discount: 11,
-		brand: 'Acer',
-		rating: 4,
-		reviews: 205,
-		stock: 7,
-		isLimitedTime: true
-	},
-	{
-		name: 'PC Gamer AMD Ryzen 5 5600X',
-		category: 'PC GAMING',
-		price: 3200,
-		originalPrice: 4000,
-		image: '/productos/headofert1.png?height=300&width=300',
-		discount: 20,
-		brand: 'AMD',
-		rating: 5,
-		reviews: 150,
-		stock: 4,
-		isLimitedTime: true
-	},
-	{
-		name: 'Silla Gaming DXRacer Racing Series',
-		category: 'SILLAS GAMING',
-		price: 399,
-		originalPrice: 499,
-		image: '/productos/sillas/silla3.png?height=300&width=300',
-		discount: 20,
-		brand: 'DXRacer',
-		rating: 4,
-		reviews: 89,
-		stock: 10,
-		isLimitedTime: true
-	},
-	{
-		name: 'Silla Gaming DXRacer Formula Series',
-		category: 'SILLAS GAMING',
-		price: 79,
-		originalPrice: 99,
-		image: '/productos/sillas/silla1.png?height=300&width=300',
-		discount: 20,
-		brand: 'DXRacer',
-		rating: 4,
-		reviews: 64,
-		stock: 12,
-		isLimitedTime: true
-	}
-]
-
-export const recommendedProducts = [
-	{
-		name: 'Smartphone Xiaomi Redmi Note 12 Pro',
-		category: 'SMARTPHONES',
-		price: 299,
-		originalPrice: 349,
-		image: '/productos/audifonos/audifono1.png?height=150&width=150',
-		discount: 14,
-		brand: 'Xiaomi',
-		rating: 4,
-		reviews: 120,
-		stock: 10,
-		isLimitedTime: false
-	},
-	{
-		name: 'Carga Inalámbrica Belkin BoostCharge',
-		category: 'ACCESORIOS',
-		price: 199,
-		originalPrice: 249,
-		image: '/productos/audifonos/audifono2.png?height=150&width=150',
-		discount: 20,
-		brand: 'Belkin',
-		rating: 4,
-		reviews: 45,
-		stock: 15,
-		isLimitedTime: true
-	},
-	{
-		name: 'Laptop HP Victus Gaming 15.6',
-		category: 'LAPTOPS',
-		price: 899,
-		originalPrice: 999,
-		image: '/productos/audifonos/audifono3.png?height=150&width=150',
-		discount: 10,
-		brand: 'HP',
-		rating: 5,
-		reviews: 78,
-		stock: 7,
-		isLimitedTime: true
-	},
-	{
-		name: 'Juego de 4K Resident Evil 4',
-		category: 'VIDEOJUEGOS',
-		price: 1099,
-		originalPrice: 1299,
-		image: '/productos/audifonos/audifono4.png?height=150&width=150',
-		discount: 15,
-		brand: 'Capcom',
-		rating: 5,
-		reviews: 321,
-		stock: 20,
-		isLimitedTime: false
-	},
-	{
-		name: 'Mochila Targus CityGear',
-		category: 'MOCHILAS',
-		price: 49,
-		originalPrice: 59,
-		image: '/productos/audifonos/audifono5.png?height=150&width=150',
-		discount: 17,
-		brand: 'Targus',
-		rating: 4,
-		reviews: 60,
-		stock: 25,
-		isLimitedTime: false
-	},
-	{
-		name: 'Mochila Targus CityGear',
-		category: 'MOCHILAS',
-		price: 39,
-		originalPrice: 59,
-		image: '/productos/audifonos/audifono6.png?height=150&width=150',
-		discount: 34,
-		brand: 'Targus',
-		rating: 4,
-		reviews: 40,
-		stock: 18,
-		isLimitedTime: true
-	}
-]
