+++ src/pages/index.astro
@@ -7,7 +7,7 @@ import { Header } from '@/components/header'
 import Footer from '@/components/footer.astro'
 import { HeroSlider } from '@/components/home/HeroSlider'
 import FiestasPatriasOfertas from '@/components/home/fiestas-patrias-ofertas.astro'
-import { products, recommendedProducts } from '@/constants/db'
+import { products, recommendedProducts } from '@/constants/db_product'
 ---
 
 <Layout>
@@ -98,7 +98,7 @@ import { products, recommendedProducts } from '@/constants/db'
 			<div class="container mx-auto px-4">
 				<div class="max-w-4xl mx-auto text-center">
 					<div class="mb-8">
-						<img src="/public/inicio/team.png?height=200&width=200" alt="Team" width={200} height={200} class="mx-auto" />
+						<img src="../inicio/team.png?height=200&width=200" alt="Team" width={200} height={200} class="mx-auto" />
 					</div>
 					<h2 class="text-3xl font-bold mb-6">¿Quienes Somos?</h2>
 					<p class="text-gray-600 leading-relaxed mb-8">
